requirement.txt
gradio
langchain-openai
youtube-transcript-api
langchain-community
langchain-core
python-dotenv